+++ examples/node/spec/helpers/reporter.js
@@ -3,6 +3,9 @@ const SpecReporter = require('jasmine-spec-reporter').SpecReporter;
 jasmine.getEnv().clearReporters();               // remove default reporter logs
 jasmine.getEnv().addReporter(new SpecReporter({  // add jasmine-spec-reporter
   spec: {
-    displayPending: true
+    displayPending: true,
+  },
+  summary: {
+    displayDuration: false,
   }
 }));