+++ examples/protractor/protractor.conf.js
@@ -1,4 +1,6 @@
 let SpecReporter = require('jasmine-spec-reporter').SpecReporter;
+// https://github.com/angular/protractor/issues/1451
+require('protractor/built/logger').Logger.logLevel = 1;
 
 exports.config = {
   framework: 'jasmine2',
@@ -22,6 +24,9 @@ exports.config = {
     jasmine.getEnv().addReporter(new SpecReporter({
       spec: {
         displayStacktrace: true
+      },
+      summary: {
+        displayDuration: false
       }
     }));
   }